backend/requirements.txt
# File Processing Dependencies
PyPDF2==3.0.1
python-docx==1.1.0
pathlib

# Embeddings & Vector Database Dependencies
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.9.1

# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
loguru==0.7.2
